fix/RSS-ECOMM-3_14: fix issues after code-review
