+++ src/core/validation/user-profile/personal-information.schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import { birthDateSchema, firstNameSchema, lastNameSchema } from '../user-registration/user-registration.schema';
+
+export const personalInformationSchema = z.object({
+  firstName: firstNameSchema,
+  lastName: lastNameSchema,
+  birthDate: birthDateSchema,
+});
