+++ src/models/index.ts
@@ -1,2 +1,8 @@
 export { type SelectOption } from './select-option.model';
-export { type RegistrationForm, type RegistrationFormAddress, type LoginForm } from './forms.model';
+export {
+  type RegistrationForm,
+  type RegistrationFormAddress,
+  type LoginForm,
+  type AddressInformationForm,
+  type PersonalInformationForm,
+} from './forms.model';
